fix(autocomplete): skip autocomplete hooks on windows

+++ packages/autocomplete/src/hooks/recache.ts
@@ -8,6 +8,8 @@ import {updateCache} from '../cache'
 import acCreate from '../commands/autocomplete/create'
 
 export const completions: Hook<any> = async function ({type, app}: {type?: 'app' | 'addon' | 'config' | 'login' | 'logout', app?: string}) {
+  // autocomplete is now in core, skip windows
+  if (this.config.windows) return
   const logInOut = type === 'login' || type === 'logout'
   const completionsDir = path.join(this.config.cacheDir, 'autocomplete', 'completions')
   const rm = () => fs.emptyDir(completionsDir)
